Вносит использование cn. Переносит типы моков в отдельный раздел. Убирает интерполяцию в строках. Изменяет наименование сеттера в карточке объявления для главной страницы. Добавляет обнуление стейта наведения на карточку для перечня на главной странице. Избавляется от магических чисел в компоненте рейтинга. Добавляет возможность использования различных систем исчисления, как 5 так и 10и бальной системы. Выносит Бэйдж премиум в отдельный компонент. Заменяет в карточке офера главной страницы применение тернарного оператора на conditional rendering. Заменяет в карточке ссылку на link для переадресации на страницу офера для данной карточки. Выносит кнопку закладки картоки офера в отдельный компонент. Выносит карту отдельным компонентом. Применяет лайаут более сегментированного использования компонентов. Вводит применения generatePath для путей карточек оферов. Производит дополнительное сегментирование табов, сортировки, формы и избранного.
