+++ markup/offer.html
@@ -167,12 +167,12 @@
                 </div>
               </div>
               <section class="offer__reviews reviews">
-                <h2 class="reviews__title">Reviews &middot; <span class="reviews__amount">1</span></h2>
+                <h2 class="reviews__title">ReviewsForm &middot; <span class="reviews__amount">1</span></h2>
                 <ul class="reviews__list">
                   <li class="reviews__item">
                     <div class="reviews__user user">
                       <div class="reviews__avatar-wrapper user__avatar-wrapper">
-                        <img class="reviews__avatar user__avatar" src="img/avatar-max.jpg" width="54" height="54" alt="Reviews avatar">
+                        <img class="reviews__avatar user__avatar" src="img/avatar-max.jpg" width="54" height="54" alt="ReviewsForm avatar">
                       </div>
                       <span class="reviews__user-name">
                         Max
