+++ src/components/favorites-card/favorites-card.tsx
@@ -1,42 +1,53 @@
 import {OfferFromList} from 'types/offer.ts';
 import Rating from 'components/rating/rating.tsx';
 import BadgePremium from 'components/badge-premium/badge-premium.tsx';
-import ButtonBookmark from 'components/buttons/button-bookmark/button-bookmark.tsx';
 import {Link} from 'react-router-dom';
 import {generatePath} from 'react-router-dom';
 import {AppRoute} from 'const.ts';
 
-type PlaceCardProps={
+type FavoritesCardProps={
   offer: OfferFromList;
-  onMouseToCard: () => void;
-  onMouseLeaveCard: () => void;
 }
 
-function PlaceCard({offer, onMouseToCard, onMouseLeaveCard}: PlaceCardProps): JSX.Element {
+function FavoritesCard({offer}: FavoritesCardProps): JSX.Element {
   return (
-    <article className="cities__card place-card" onMouseOver={onMouseToCard} onMouseLeave={onMouseLeaveCard}>
+    <article className="favorites__card place-card">
       {offer.isPremium && <BadgePremium blockName='place-card' />}
-      <div className="cities__image-wrapper place-card__image-wrapper">
+      <div className="favorites__image-wrapper place-card__image-wrapper">
         <a href="#">
           <img
             className="place-card__image"
             src={offer.previewImage}
-            width={260}
-            height={200}
+            width={150}
+            height={110}
             alt={offer.title}
           />
         </a>
       </div>
-      <div className="place-card__info">
+      <div className="favorites__card-info place-card__info">
         <div className="place-card__price-wrapper">
           <div className="place-card__price">
             <b className="place-card__price-value">€{offer.price}</b>
-            <span className="place-card__price-text">/&nbsp;night</span>
+            <span className="place-card__price-text">
+                            /&nbsp;night
+            </span>
           </div>
-          <ButtonBookmark isFavorite={offer.isFavorite}/>
+          <button
+            className="place-card__bookmark-button place-card__bookmark-button--active button"
+            type="button"
+          >
+            <svg
+              className="place-card__bookmark-icon"
+              width={18}
+              height={19}
+            >
+              <use xlinkHref="#icon-bookmark"/>
+            </svg>
+            <span className="visually-hidden">In bookmarks</span>
+          </button>
         </div>
         <Rating rating={offer.rating} calculusSystem={5} />
-        <h2 className="place-card__name">`
+        <h2 className="place-card__name">
           <Link to={generatePath(AppRoute.Offer, {id: offer.id})}>
             {offer.title}
           </Link>
@@ -47,4 +58,4 @@ function PlaceCard({offer, onMouseToCard, onMouseLeaveCard}: PlaceCardProps): JS
   );
 }
 
-export default PlaceCard;
+export default FavoritesCard;
