+++ src/components/header/header.tsx
@@ -6,7 +6,7 @@ function Header(): JSX.Element{
       <div className="container">
         <div className="header__wrapper">
           <div className="header__left">
-            <Logo blockName={'header'} active/>
+            <Logo blockName="header" active/>
           </div>
           <nav className="header__nav">
             <ul className="header__nav-list">
