+++ src/components/offers-list/offers-list.tsx
@@ -1,8 +1,8 @@
-import {OfferFromList} from 'mocks/offers.ts';
+import {OfferFromList} from 'types/offer.ts';
 import PlaceCard from '../place-card/place-card.tsx';
 import {useState} from 'react';
 
-type OffersListProps = {
+export type OffersListProps = {
   offers: OfferFromList[];
 }
 
@@ -10,10 +10,10 @@ function OffersList({offers}: OffersListProps): JSX.Element {
   const [, setCard] = useState<string | null>(null);
 
   return (
-    <div className="cities__places-list places__list tabs__content">
+    <div className="cities__places-list places__list">
       {
         offers.map((item) =>
-          <PlaceCard key={item.id} offer={item} setCard={() => setCard(item.id)}/>
+          <PlaceCard key={item.id} offer={item} onMouseToCard={() => setCard(item.id)} onMouseLeaveCard={() => setCard(null)}/>
         )
       }
     </div>
