+++ src/pages/not-found/not-found.tsx
@@ -1,12 +1,16 @@
 import {Link} from 'react-router-dom';
 import styles from './not-found.module.css';
+import Layout from 'components/layout/layout.tsx';
+import Header from 'components/header/header.tsx';
 
 function NotFound(): JSX.Element {
   return (
-    <div className={styles.container}>
-      <h1>404 Not Found</h1>
-      <Link to="/" className={styles.link}>Вернуться на главную</Link>
-    </div>
+    <Layout header={<Header/>}>
+      <div className={styles.container}>
+        <h1>404 Not Found</h1>
+        <Link to="/" className={styles.link}>Вернуться на главную</Link>
+      </div>
+    </Layout>
   );
 }
