+++ src/types/offer.ts
@@ -1,0 +1,47 @@
+export type Offer={
+  id: string;
+  title: string;
+  type: string;
+  price: number;
+  city: OfferCity;
+  location: OfferLocation;
+  isFavorite: boolean;
+  isPremium: boolean;
+  rating: number;
+  description: string;
+  bedrooms: number;
+  goods: string[];
+  host: OfferHost;
+  images: string[];
+  maxAdults: number;
+}
+
+type OfferCity={
+  name: string;
+  location: OfferLocation;
+}
+
+type OfferLocation={
+  latitude: number;
+  longitude: number;
+  zoom: number;
+}
+
+type OfferHost={
+  name: string;
+  avatarUrl: string;
+  isPro: boolean;
+}
+
+export type OfferFromList={
+  id: string;
+  title: string;
+  type: string;
+  price: number;
+  city: OfferCity;
+  location: OfferLocation;
+  isFavorite: boolean;
+  isPremium: boolean;
+  rating: number;
+  previewImage: string;
+}
