+++ src/types/rewiew.ts
@@ -1,0 +1,13 @@
+type ReviewUser={
+  name: string;
+  avatarUrl: string;
+  isPro: boolean;
+}
+
+export type Review={
+  id: string;
+  date: string;
+  user: ReviewUser;
+  comment: string;
+  rating: number;
+}
